+++ htdocs/theme/oblyon/main_menu_fa_icons.inc.php
@@ -420,3 +420,68 @@ div.mainmenu.tools::before {
 div.mainmenu.website::before {
     content: "\f542";
 }
+
+/* Define color of some picto */
+
+.fa-phone, .fa-mobile-alt, .fa-fax {
+    opacity: 0.7;
+    color: #440;
+}
+.fa-at, .fa-external-link-alt, .fa-share-alt {
+    opacity: 0.7;
+    color: #304;
+}
+.fa-trash {
+    color: #666;
+}
+.fa-trash:hover:before {
+    color: #800;
+}
+.fa-play {
+    color: #444;
+}
+.fa-link, .fa-unlink {
+    color: #555;
+}
+
+/* Define square Dolibarr logo in pure CSS */
+
+.fa-dolibarr-css{
+    color: #235481;
+    background: currentColor;
+    height: 150px;
+    width: 150px;
+    position: relative;
+}
+.fa-dolibarr-css:before{
+    content: '';
+    position: absolute;
+    left: 19%;
+    top: 17%;
+    width: 25%;
+    height: 25%;
+    border: solid 30px white;
+    border-radius: 0% 200% 200% 0% / 0% 180% 180% 0%;
+}
+.fa-dolibarr-css:after{
+    content: '';
+    position: absolute;
+    left: 19%;
+    top: 17%;
+    width: 5px;
+    height: 25%;
+    border-bottom: solid 60px currentColor;
+    margin-left: 30px;
+}
+
+.em092 {
+    font-size: 0.92em;
+}
+
+.em088 {
+    font-size: 0.88em;
+}
+
+.em080 {
+    font-size: 0.80em;
+}